requirements.txt
langchain==0.3.26
langchain-community==0.3.26
langchain-google-genai
langchain-pinecone==0.2.0
python-dotenv==1.1.0
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
pydantic>=2,<3
-e .